requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
Cython==0.29.35
docopt==0.6.2
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
pipreqs==0.4.13
PyOpenGL @ file:///C:/Program%20Files%20%28x86%29/ZED%20SDK/PyOpenGL-3.1.6-cp310-cp310-win_amd64.whl
PyOpenGL-accelerate @ file:///C:/Program%20Files%20%28x86%29/ZED%20SDK/PyOpenGL_accelerate-3.1.6-cp310-cp310-win_amd64.whl
pyzed @ file:///C:/Program%20Files%20%28x86%29/ZED%20SDK/pyzed-4.0-cp310-cp310-win_amd64.whl
requests==2.31.0
urllib3==2.0.3
yarg==0.1.9
matplotlib-3.7.2
